# perf: short circuit Oid.BOOL in getBoolean