@davecramer 2017-02-08
Just for my elucidation is the score nanoseconds, milliseconds, or ?

Dave Cramer

On 8 February 2017 at 12:38, Jorge Solorzano <notifications@github.com>
wrote:

> Does this buy much?
>
> Preliminary test (added to the PR) with the length check as suggested:
> Before:
>
> Benchmark                  (rowsize)  Mode  Cnt        Score       Error  Units
> ProcessBoolean.getBoolean          5  avgt   10      690.743 ±    21.417  ns/op
> ProcessBoolean.getBoolean         10  avgt   10     1623.945 ±    50.107  ns/op
> ProcessBoolean.getBoolean         50  avgt   10     9092.337 ±   149.268  ns/op
> ProcessBoolean.getBoolean        100  avgt   10    17716.010 ±   913.568  ns/op
> ProcessBoolean.getBoolean      10000  avgt   10  1788860.346 ± 14054.267  ns/op
>
> After:
>
> Benchmark                  (rowsize)  Mode  Cnt      Score     Error  Units
> ProcessBoolean.getBoolean          5  avgt   10     12.155 ±   0.041  ns/op
> ProcessBoolean.getBoolean         10  avgt   10     20.835 ±   0.096  ns/op
> ProcessBoolean.getBoolean         50  avgt   10    103.061 ±   2.483  ns/op
> ProcessBoolean.getBoolean        100  avgt   10    194.745 ±   0.969  ns/op
> ProcessBoolean.getBoolean      10000  avgt   10  21071.552 ± 170.534  ns/op
>
> That is about ~95% performance increase.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/pgjdbc/pgjdbc/pull/745#issuecomment-278402949>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAYz9ij_a93xbTFjXRMWYeiKieo6jH-Dks5raf2tgaJpZM4L6KE->
> .
>