@jorsol 2017-02-08
@davecramer it's nanoseconds