@vlsi 2017-02-09
I've tested the PR a bit. I was able to activate logging via `-DloggerLevel=DEBUG` (`loggerFile` was inherited).

However I do not like log format much.
Here's what I get:
```
фев 10, 2017 12:21:50 AM org.postgresql.Driver connect
FINE: Connecting with URL: jdbc:postgresql://localhost:5432/postgres?ApplicationName=Driver Tests&loggerLevel=DEBUG&loggerFile=target/pgjdbc-tests.log
фев 10, 2017 12:21:50 AM org.postgresql.jdbc.PgConnection <init>
FINE: PostgreSQL JDBC Driver /*$mvn.project.property.parsedversion.osgiversion$*/
фев 10, 2017 12:21:50 AM org.postgresql.jdbc.PgConnection setDefaultFetchSize
FINE:   setDefaultFetchSize = 0
фев 10, 2017 12:21:50 AM org.postgresql.jdbc.PgConnection setPrepareThreshold
FINE:   setPrepareThreshold = 5
фев 10, 2017 12:21:50 AM org.postgresql.core.v3.ConnectionFactoryImpl openConnectionImpl
FINE: Trying to establish a protocol version 3 connection to localhost:5432
фев 10, 2017 12:21:50 AM org.postgresql.core.v3.ConnectionFactoryImpl openConnectionImpl
FINE: Receive Buffer Size is 408 300
```

1) two-line log records look odd to me. Baiscally the first line is date+location, and the second line is the message. I would prefer to have single log lines. Any thoughts here?
2) "automatic" class and method names (see the line with timestamps). Of course they are not free, are they?

I've did a bit of google and a bit of benchmarks. Here's what I've got:

https://bugs.openjdk.java.net/browse/JDK-6511515: poor performance of LogRecord.inferCaller depending on java.lang.Throwable.getStackTraceElement

```java
package org.postgresql.benchmark.connection;

import org.openjdk.jmh.annotations.Benchmark;
import org.openjdk.jmh.annotations.BenchmarkMode;
import org.openjdk.jmh.annotations.Fork;
import org.openjdk.jmh.annotations.Measurement;
import org.openjdk.jmh.annotations.Mode;
import org.openjdk.jmh.annotations.OutputTimeUnit;
import org.openjdk.jmh.annotations.Scope;
import org.openjdk.jmh.annotations.State;
import org.openjdk.jmh.annotations.Threads;
import org.openjdk.jmh.annotations.Warmup;
import org.openjdk.jmh.infra.Blackhole;
import org.openjdk.jmh.profile.GCProfiler;
import org.openjdk.jmh.runner.Runner;
import org.openjdk.jmh.runner.RunnerException;
import org.openjdk.jmh.runner.options.Options;
import org.openjdk.jmh.runner.options.OptionsBuilder;

import java.util.concurrent.TimeUnit;
import java.util.logging.Level;
import java.util.logging.LogRecord;
import java.util.logging.SimpleFormatter;

@Fork(1)
@Measurement(iterations = 5)
@Warmup(iterations = 5)
@State(Scope.Thread)
@Threads(1)
@BenchmarkMode(Mode.AverageTime)
@OutputTimeUnit(TimeUnit.MICROSECONDS)
public class LoggingBenchmark {

  String msg = "test";
  SimpleFormatter formatter = new SimpleFormatter();

  @Benchmark
  public LogRecord baseline() {
    return new LogRecord(Level.FINE, "test");
  }

  @Benchmark
  public LogRecord locationInfo(Blackhole bh) {
    LogRecord log = new LogRecord(Level.FINE, "test");
    bh.consume(log.getSourceClassName());
    bh.consume(log.getSourceMethodName());
    return log;
  }

  @Benchmark
  public String format(Blackhole bh) {
    LogRecord log = new LogRecord(Level.FINE, "test");
    return formatter.format(log);
  }


  public static void main(String[] args) throws RunnerException {
    Options opt = new OptionsBuilder()
        .include(LoggingBenchmark.class.getSimpleName())
        .addProfiler(GCProfiler.class)
        .detectJvmArgs()
        .build();

    new Runner(opt).run();
  }

}
```

Results:

```
Benchmark                                    Mode  Cnt      Score      Error   Units
LoggingBenchmark.baseline                    avgt    5      0,084 ±    0,003   us/op
LoggingBenchmark.baseline:·gc.alloc.rate     avgt    5    818,191 ±   30,430  MB/sec
LoggingBenchmark.locationInfo                avgt    5     21,710 ±    3,459   us/op
LoggingBenchmark.locationInfo:·gc.alloc.rate avgt    5     47,141 ±    7,540  MB/sec
LoggingBenchmark.format                      avgt    5     36,515 ±    5,391   us/op
LoggingBenchmark.format:·gc.alloc.rate       avgt    5    338,382 ±   48,791  MB/sec
```

What it means is it takes 36 microseconds to format a single log line with a simple "test" message.
The main time consumer is indeed location calculation that takes like 22us.

Just in case: simple JDBC query (prepare+execute+fetch) over localhost connection takes comparable time.

The formatting is performed under synchronization (`SimpleFormatter#format` is `synchronized`), so I don't even bother to benchmark how it would break in multi-threaded environment.

I'm not saying this logging facility should always be enabled in production at DEBUG/TRACE levels, however this level of overhead might bite us (or pgjdbc users :) ).

Any thoughts if we should cleanup log format before merging the PR in?