@jorsol 2017-02-10
More benchmark
Before:
```
Benchmark                  (rowsize)  (tokens)  Mode  Cnt        Score       Error  Units
ProcessBoolean.getBoolean          5         0  avgt   10      716.504 ±     8.007  ns/op
ProcessBoolean.getBoolean          5        10  avgt   10      668.592 ±     5.034  ns/op
ProcessBoolean.getBoolean          5        50  avgt   10      765.153 ±     8.295  ns/op
ProcessBoolean.getBoolean          5       100  avgt   10      859.289 ±     8.129  ns/op
ProcessBoolean.getBoolean         10         0  avgt   10     1567.071 ±     8.863  ns/op
ProcessBoolean.getBoolean         10        10  avgt   10     1582.278 ±    13.387  ns/op
ProcessBoolean.getBoolean         10        50  avgt   10     1704.858 ±    78.937  ns/op
ProcessBoolean.getBoolean         10       100  avgt   10     1825.133 ±    64.091  ns/op
ProcessBoolean.getBoolean         50         0  avgt   10     8884.121 ±   302.890  ns/op
ProcessBoolean.getBoolean         50        10  avgt   10     8764.162 ±   308.606  ns/op
ProcessBoolean.getBoolean         50        50  avgt   10     8781.140 ±   250.280  ns/op
ProcessBoolean.getBoolean         50       100  avgt   10     9031.006 ±   300.511  ns/op
ProcessBoolean.getBoolean        100         0  avgt   10    18512.088 ±   941.861  ns/op
ProcessBoolean.getBoolean        100        10  avgt   10    17759.530 ±   502.616  ns/op
ProcessBoolean.getBoolean        100        50  avgt   10    17987.266 ±   518.799  ns/op
ProcessBoolean.getBoolean        100       100  avgt   10    18534.797 ±   518.986  ns/op
ProcessBoolean.getBoolean      10000         0  avgt   10  1762470.368 ± 33905.257  ns/op
ProcessBoolean.getBoolean      10000        10  avgt   10  1806601.186 ± 99918.977  ns/op
ProcessBoolean.getBoolean      10000        50  avgt   10  1807746.782 ± 34044.221  ns/op
ProcessBoolean.getBoolean      10000       100  avgt   10  1763167.027 ± 27145.896  ns/op
```

After:
```
Benchmark                  (rowsize)  (tokens)  Mode  Cnt      Score      Error  Units
ProcessBoolean.getBoolean          5         0  avgt   10     33.610 ±    0.191  ns/op
ProcessBoolean.getBoolean          5        10  avgt   10     41.735 ±    0.075  ns/op
ProcessBoolean.getBoolean          5        50  avgt   10    116.491 ±    1.377  ns/op
ProcessBoolean.getBoolean          5       100  avgt   10    208.021 ±    1.434  ns/op
ProcessBoolean.getBoolean         10         0  avgt   10     72.764 ±    3.320  ns/op
ProcessBoolean.getBoolean         10        10  avgt   10     79.348 ±    0.609  ns/op
ProcessBoolean.getBoolean         10        50  avgt   10    153.837 ±    0.534  ns/op
ProcessBoolean.getBoolean         10       100  avgt   10    247.109 ±    3.211  ns/op
ProcessBoolean.getBoolean         50         0  avgt   10    360.975 ±    4.420  ns/op
ProcessBoolean.getBoolean         50        10  avgt   10    366.895 ±    2.675  ns/op
ProcessBoolean.getBoolean         50        50  avgt   10    437.836 ±    0.907  ns/op
ProcessBoolean.getBoolean         50       100  avgt   10    529.986 ±    0.840  ns/op
ProcessBoolean.getBoolean        100         0  avgt   10    708.309 ±    2.597  ns/op
ProcessBoolean.getBoolean        100        10  avgt   10    735.806 ±   18.164  ns/op
ProcessBoolean.getBoolean        100        50  avgt   10    797.098 ±   10.979  ns/op
ProcessBoolean.getBoolean        100       100  avgt   10    892.736 ±   18.678  ns/op
ProcessBoolean.getBoolean      10000         0  avgt   10  73050.736 ± 2890.068  ns/op
ProcessBoolean.getBoolean      10000        10  avgt   10  73380.023 ± 2193.849  ns/op
ProcessBoolean.getBoolean      10000        50  avgt   10  74308.542 ± 3546.948  ns/op
ProcessBoolean.getBoolean      10000       100  avgt   10  73160.013 ± 2428.678  ns/op
```